a bunch of changes

+++ fw/src/util.h
@@ -6,12 +6,12 @@
 typedef uint8_t u8;
 typedef uint16_t u16;
 typedef uint32_t u32;
-typedef uint32_t u64;
+typedef uint64_t u64;
 
 typedef int8_t i8;
 typedef int16_t i16;
 typedef int32_t i32;
-typedef int32_t i64;
+typedef int64_t i64;
 
 #define LIKELY(x) __builtin_expect(static_cast<bool>(x), 1)
 #define UNLIKELY(x) __builtin_expect(static_cast<bool>(x), 0)
